+++ GameInstance.h
@@ -11,7 +11,7 @@
 class GameInstance
 {
 private:
-	int accIndex;
+	size_t accIndex;
 	std::vector<std::string>* database;
 	uint64_t numCookies;
 	uint64_t cookieRate;
@@ -27,21 +27,21 @@ private:
 	std::string nickname;
 public:
 	bool bDestroyed;
-	GameInstance(int _accIndex,std::vector<std::string>* _database,bool _bIsGameMaster=false,std::string _nickname="");
+	GameInstance(size_t _accIndex,std::vector<std::string>* _database,bool _bIsGameMaster=false,std::string _nickname="");
 	void addCookie(uint64_t amountToAdd);
-	bool subtractCookie(double amountToSubtract);
+	bool subtractCookie(uint64_t amountToSubtract);
 	void saveChanges();
 	void addToRate(uint64_t amountToAdd);
 	void addToClickRate(uint64_t amountToAdd);
-	double getNumCookies(); // get cookies of logged in user
-	double getNumCookies(std::string user);
-	double getCookieRate();
-	double getCookieClickRate();
+	uint64_t getNumCookies(); // get cookies of logged in user
+	uint64_t getNumCookies(std::string user);
+	uint64_t getCookieRate();
+	uint64_t getCookieClickRate();
 	std::vector<std::string>* getFollowing();
 	std::vector<int>* getAccIndices();
 	bool destroyed();
 	std::string getUserName(int index);
-	int getAccIndex();
+	size_t getAccIndex();
 	void follow(std::string &choice,std::vector<std::string>* choices);
 	void refreshFollowingList();
 	bool isGameMaster();
